pyproject.toml
[project]
name = "django4-todo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "drf-yasg>=1.21.10",
]

[dependency-groups]
tests = [
    "coverage>=7.8.0",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "todo.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]